+++ lib/src/app/routes.dart
@@ -1,17 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:jdr_maker/src/app/controllers/navigation_controller.dart';
 import 'package:jdr_maker/src/app/views/accueil_view.dart';
-import 'package:jdr_maker/src/app/views/connexion_view.dart';
+import 'package:jdr_maker/src/app/views/connexion/connexion_view.dart';
+import 'package:jdr_maker/src/app/views/connexion/inscription_view.dart';
 import 'package:jdr_maker/src/app/views/creerJDR/debut_jdr_view.dart';
 import 'package:jdr_maker/src/app/views/evenements/evenement_create_view.dart';
 import 'package:jdr_maker/src/app/views/evenements/evenement_edit_view.dart';
 import 'package:jdr_maker/src/app/views/evenements/evenement_view.dart';
 import 'package:jdr_maker/src/app/views/evenements/evenements_view.dart';
-import 'package:jdr_maker/src/app/views/inscription_view.dart';
 import 'package:jdr_maker/src/app/views/lieux/lieu_create.dart';
 import 'package:jdr_maker/src/app/views/lieux/lieu_edit.dart';
 import 'package:jdr_maker/src/app/views/lieux/lieu_view.dart';
 import 'package:jdr_maker/src/app/views/lieux/lieux_view.dart';
+import 'package:jdr_maker/src/app/views/modifier_projet/modifier_projet_view.dart';
 import 'package:jdr_maker/src/app/views/objets/objet_create.dart';
 import 'package:jdr_maker/src/app/views/objets/objet_edit.dart';
 import 'package:jdr_maker/src/app/views/objets/objet_view.dart';
@@ -41,7 +42,7 @@ List<Page> applicationRoutes(context) {
   List<Page> liste = [];
 
   // Route par défaut
-  liste.add(MaterialPage(child: AccueilView()));
+  liste.add(MaterialPage(child: ConnexionView()));
 
   // Routing
   // Toujours un / en début de route
@@ -147,6 +148,13 @@ List<Page> applicationRoutes(context) {
     // =======================================================
     case "/lance":
       liste.add(MaterialPage(child: LanceDes()));
+      break;
+
+    // =======================================================
+    // Routes liés aux options
+    // =======================================================
+    case "/modifier_projet":
+      liste.add(MaterialPage(child: ModifierProjetView()));
       break;
   }